Nginx image for index + tar.gz
Assets: flat structure, use version in name
Non-release artifacts in ci/build directory

Current github releases use a URL like this:
{{EXTERNAL_URL}}/swift-v0.1.4/incubator.swift.templates.simple.tar.gz

Kabanero uses versions in their archives, so it looks more like this:
{{EXTERNAL_URL}}/incubator.java-spring-boot2.v0.3.6.templates.default.tar.gz

./ci/package.sh has been updated to deal with existing template
archives with and without the version in the filename.
Archives that are built (either explicitly, or because they are
missing), will be created with the version in the filename.

./ci/prefetch.sh will download all of the artifacts referenced by
index yaml files listed (space separated) in PREFETCH_INDEX_YAML.

Signed-off-by: Erin Schnabel <schnabel@us.ibm.com>