swift: override Docker labels from base image (#470)
The base `swift:5.1` Docker image defines two labels: `Description`
and `maintainer`. These were being propagated through to the
`appsody/swift` images, which was misleading as the wrong maintainer was
listed in our image.

Override these labels in the Swift stack's `Dockerfile-stack` so that
the correct information appears in `docker describe` output.