java-spring-boot2: image layers, common base image
Refactor the final Dockerfile to take better advantage of
image layer caching when compiling and building the final
image.

Use a value from the template to specify the base image for
both the stack and the compile stage of the Dockerfile.
Otherwise ensure that the stack image and the compile stage
of the build image are the same (same initialization steps).

Assign a label to the compile image specifying the stack
id and version, the image is otherwise unidentified, though
though it is in the cache.

Signed-off-by: Erin Schnabel <schnabel@us.ibm.com>